+++ tools/webpack.config.js
@@ -4,7 +4,7 @@ const autoprefixer = require('autoprefixer')
 const { BannerPlugin, DefinePlugin } = require('webpack')
 const CompressionPlugin = require('compression-webpack-plugin')
 const BundleAnalyzerPlugin = require('webpack-bundle-analyzer').BundleAnalyzerPlugin
-const CleanWebpackPlugin = require('clean-webpack-plugin')
+const { CleanWebpackPlugin } = require('clean-webpack-plugin')
 const HtmlWebpackPlugin = require('html-webpack-plugin')
 const HtmlWebpackHarddiskPlugin = require('html-webpack-harddisk-plugin')
 const langFiles = require('formbuilder-languages')
@@ -71,25 +71,21 @@ const webpackConfig = {
           {
             loader: 'style-loader',
             options: {
-              attrs: {
+              attributes: {
                 class: 'formBuilder-injected-style',
               },
-              sourceMap: !PRODUCTION,
             },
           },
           {
             loader: 'css-loader',
             options: {
-              camelCase: true,
               sourceMap: !PRODUCTION,
             },
           },
           {
             loader: 'postcss-loader',
             options: {
-              plugins: [
-                autoprefixer(),
-              ],
+              plugins: [autoprefixer()],
               sourceMap: !PRODUCTION,
             },
           },
@@ -150,6 +146,7 @@ const webpackConfig = {
     inline: true,
     contentBase: 'demo/',
     noInfo: true,
+    open: true
   },
 }